+++ src/components/cart/CartSidebar/CartSidebar.tsx
@@ -26,22 +26,7 @@ function CartSidebar() {
           ) : (
             <>
               {items.map((item) => (
-                <CartItem
-                  key={item.id}
-                  item={{
-                    ...item,
-                    name: 'name',
-                    brand: 'brand',
-                    facets: [
-                      { key: 'color', value: 'todo' },
-                      { key: 'size', value: 'todo' },
-                    ],
-                    image: {
-                      src: 'https://fashioneurope.vtexassets.com/arquivos/ids/155504-192-auto',
-                      alt: 'todo',
-                    },
-                  }}
-                />
+                <CartItem key={item.id} item={item as any} />
               ))}
             </>
           )}
